requirements.txt
numpy
scipy
pandas
scikit-learn
matplotlib
tqdm
rich
pytorch-lightning
lightning
omegaconf
wandb
einops
opt_einsum
cmake # For pykeops support
# pykeops # Only for S4D. If there are installation problems with pykeops==2.x, try pykeops==1.5
transformers==4.26.1 # For some schedulers and tokenizers
torchvision
timm
prettytable
numerize
git-lfs
seaborn
jsonargparse[signatures]>=4.27.7
# hyperparameter optimization
optuna
optuna-integration

# Dataset specific packages
torchtext==0.14.0 # this needs to align with the pytorch version
datasets # LRA
